+++ src/my_app/lib/core/domain/bloc/hidable/hidable_bloc.dart
@@ -1,0 +1,10 @@
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:injectable/injectable.dart';
+import 'package:very_good_core/app/helpers/extensions/cubit_ext.dart';
+
+@lazySingleton
+class HidableBloc extends Cubit<bool> {
+  HidableBloc() : super(true);
+
+  void setVisibility({required bool isVisible}) => safeEmit(isVisible);
+}