+++ src/my_app/test/unit/core/bloc/app_core_bloc_test.dart
@@ -1,49 +1,29 @@
+import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:very_good_core/app/constants/enum.dart';
-import 'package:very_good_core/app/constants/route_name.dart';
 import 'package:very_good_core/core/domain/bloc/app_core/app_core_bloc.dart';
 
 void main() {
   late AppCoreBloc appCoreBloc;
+  late Map<AppScrollController, ScrollController> scrollControllers;
 
   setUp(() {
     appCoreBloc = AppCoreBloc();
+    scrollControllers = <AppScrollController, ScrollController>{
+      AppScrollController.home: ScrollController(debugLabel: AppScrollController.home.name),
+      AppScrollController.profile: ScrollController(debugLabel: AppScrollController.profile.name),
+    };
   });
 
-  group('AppCore initializeScrollControllers', () {
-    test(
-      'should initialize a scroll controllers',
-      () {
-        appCoreBloc.initialize();
-
-        expect(
-          appCoreBloc.state.scrollControllers[AppScrollController.home],
-          isA<ScrollController>(),
-        );
-
-        expect(
-          appCoreBloc.state.scrollControllers[AppScrollController.profile],
-          isA<ScrollController>(),
-        );
-      },
-    );
-  });
-
-  group('AppCore getScrollControllers', () {
-    test(
-      'should return a scroll controllers',
-      () {
-        expect(
-          appCoreBloc.getScrollController(RouteName.home.name),
-          isA<ScrollController>(),
-        );
-
-        expect(
-          appCoreBloc.getScrollController(RouteName.profile.name),
-          isA<ScrollController>(),
-        );
-      },
+  group('AppCore setScrollControllers', () {
+    blocTest<AppCoreBloc, AppCoreState>(
+      'should set scrollControllers',
+      build: () => appCoreBloc,
+      act: (AppCoreBloc bloc) async => bloc.setScrollControllers(scrollControllers),
+      expect: () => <AppCoreState>[
+        appCoreBloc.state.copyWith(scrollControllers: scrollControllers),
+      ],
     );
   });
 }