+++ src/my_app/test/widget/core/widgets/very_good_core_nav_bar_test.dart
@@ -5,10 +5,12 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
+import 'package:nested/nested.dart';
 import 'package:very_good_core/app/constants/enum.dart';
 import 'package:very_good_core/app/constants/route_name.dart';
 import 'package:very_good_core/app/themes/app_theme.dart';
 import 'package:very_good_core/core/domain/bloc/app_core/app_core_bloc.dart';
+import 'package:very_good_core/core/domain/bloc/hidable/hidable_bloc.dart';
 import 'package:very_good_core/core/presentation/widgets/very_good_core_nav_bar.dart';
 
 import '../../../utils/mock_go_router_provider.dart';
@@ -22,17 +24,20 @@ import 'very_good_core_nav_bar_test.mocks.dart';
   MockSpec<StatefulNavigationShell>(),
   MockSpec<RouteMatchList>(),
   MockSpec<AppCoreBloc>(),
+  MockSpec<HidableBloc>(),
 ])
 void main() {
   late MockGoRouter router;
   late MockGoRouterDelegate routerDelegate;
   late MockRouteMatchList currentConfiguration;
   late MockAppCoreBloc appCoreBloc;
+  late MockHidableBloc hidableBloc;
   late MockStatefulNavigationShell navigationShell;
   late Map<AppScrollController, ScrollController> scrollControllers;
 
   setUp(() {
     appCoreBloc = MockAppCoreBloc();
+    hidableBloc = MockHidableBloc();
     scrollControllers = mockScrollControllers;
     when(appCoreBloc.stream).thenAnswer(
       (_) => Stream<AppCoreState>.fromIterable(
@@ -42,11 +47,18 @@ void main() {
       ),
     );
     when(appCoreBloc.state).thenAnswer(
-      (_) =>
-          AppCoreState.initial().copyWith(scrollControllers: scrollControllers),
+      (_) => AppCoreState.initial().copyWith(scrollControllers: scrollControllers),
     );
-    when(appCoreBloc.getScrollController(any))
-        .thenAnswer((_) => ScrollController());
+    when(hidableBloc.stream).thenAnswer(
+      (_) => Stream<bool>.fromIterable(
+        <bool>[
+          true,
+        ],
+      ),
+    );
+    when(hidableBloc.state).thenAnswer(
+      (_) => true,
+    );
   });
 
   MockGoRouter setUpRouter(String path, int index) {
@@ -55,21 +67,26 @@ void main() {
     navigationShell = MockStatefulNavigationShell();
     currentConfiguration = MockRouteMatchList();
     when(currentConfiguration.uri).thenAnswer((_) => Uri(path: path));
-    when(routerDelegate.currentConfiguration)
-        .thenAnswer((_) => currentConfiguration);
+    when(routerDelegate.currentConfiguration).thenAnswer((_) => currentConfiguration);
     when(router.routerDelegate).thenAnswer((_) => routerDelegate);
     when(navigationShell.currentIndex).thenAnswer((_) => index);
     return router;
   }
 
-  Widget buildNavBar(MockGoRouter router) => BlocProvider<AppCoreBloc>(
-        create: (BuildContext context) => appCoreBloc,
+  Widget buildNavBar(MockGoRouter router) => MultiBlocProvider(
+        providers: <SingleChildWidget>[
+          BlocProvider<AppCoreBloc>(
+            create: (BuildContext context) => appCoreBloc,
+          ),
+          BlocProvider<HidableBloc>(
+            create: (BuildContext context) => hidableBloc,
+          ),
+        ],
         child: MockLocalization(
           child: MockGoRouterProvider(
             router: router,
             child: PreferredSize(
-              preferredSize:
-                  const Size.fromHeight(AppTheme.defaultNavBarHeight),
+              preferredSize: const Size.fromHeight(AppTheme.defaultNavBarHeight),
               child: VeryGoodCoreNavBar(navigationShell: navigationShell),
             ),
           ),