+++ src/my_app/test/widget/features/home/views/home_screen_test.dart
@@ -61,8 +61,6 @@ void main() {
     when(appCoreBloc.state).thenReturn(
       AppCoreState.initial().copyWith(scrollControllers: scrollControllers),
     );
-    when(appCoreBloc.getScrollController(any))
-        .thenAnswer((_) => ScrollController());
   });
 
   MockGoRouter setUpRouter(String path, int index) {